requirements.txt
# RiftboundOCR Service - Dependencies
# 
# IMPORTANT: Python 3.13 Compatibility Issue
# - paddleocr 2.9.1 requires numpy < 2.0
# - Python 3.13 requires numpy >= 2.1.0
# 
# RECOMMENDED: Use Python 3.11 or 3.12 for this project
# For Python 3.13: Try requirements-py313-latest.txt (experimental)
#
# Tested and working on: Python 3.11, Python 3.12

# Core OCR Engines
paddlepaddle==3.0.0
paddleocr==2.9.1
easyocr==1.7.2

# Image Processing
pillow==10.4.0
opencv-python==4.10.0.84
numpy>=1.26.4  # Python 3.13 will use numpy 2.x automatically

# Fuzzy Matching
rapidfuzz==3.10.1

# API Framework
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP Client (for API integration)
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Development
python-dotenv==1.0.1

